+++ example/README.md
@@ -1,3 +1,5 @@
 # dough_sample_app
 
 This example application demonstrates how to use the `dough` package within a simple Flutter app.
+
+## Draggable Dough
