+++ example/lib/dough_demo_pages/draggable_dough.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class DraggableDough extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Pressable Dough'),
+      ),
+      body: Center(),
+    );
+  }
+}
