+++ example/lib/dough_demo_pages/pressable_dough.dart
@@ -1,6 +1,7 @@
 import 'package:dough/dough.dart';
 import 'package:flutter/material.dart';
 
+/// This page demonstrates how to use the pressable dough widget.
 class PressableDoughPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
