+++ example/lib/routes.dart
@@ -1,13 +1,15 @@
-import 'package:dough_tester/pages/pressable_dough.dart';
+import 'main.dart';
 
-import 'pages/home.dart';
+import 'dough_demo_pages/pressable_dough.dart';
 
 class Routes {
   static final kHome = '/';
   static final kPressableDough = '/pressable-dough';
+  static final kDraggableDough = '/draggable-dough';
 
   static dynamic define() => {
         kHome: (context) => HomePage(),
         kPressableDough: (context) => PressableDoughPage(),
+        kDraggableDough: (context) => PressableDoughPage(),
       };
 }
