+++ lib/widgets/draggable_dough.dart
@@ -1,8 +1,17 @@
 part of dough;
 
-class DraggableDough extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    throw UnimplementedError();
-  }
-}
+// class DraggableDoughTarget<T> extends StatelessWidget {
+  
+// }
+
+// class DraggableDough<T> extends StatelessWidget {
+//   final Widget child;
+//   final Widget feedback;
+//   final T data;
+//   final Axis axis;
+
+//   @override
+//   Widget build(BuildContext context) {
+//     throw UnimplementedError();
+//   }
+// }
